+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Open a popover to the delivery button.
+
 ## [0.4.2] - 2024-11-06
 
 
@@ -19,7 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Drawer opens behind the sidebar
-
 
 ## [0.4.0] - 2024-10-04