+++ CHANGELOG.md
@@ -9,10 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Open a popover to the delivery button.
+- Popover to open the location drawer;
+- Popover with input to add postal code.
 
 ## [0.4.2] - 2024-11-06
-
 
 ### Fixed